changed cpu total to float instead of int

+++ pkg/core/host.go
@@ -5,7 +5,7 @@ import "encoding/json"
 type Host struct {
 	HostID      string  `json:"hostid"`
 	StateID     int     `json:"stateid"`
-	TotalCPU    int64   `json:"total_cpu"`
+	TotalCPU    float64 `json:"total_cpu"`
 	TotalMemory int64   `json:"total_memory"`
 	UsageCPU    float64 `json:"usage_cpu"`
 	UsageMemory int64   `json:"usage_memory"`
